migrate from middleware to proxy as per latest convention
